+++ ewm-service/src/main/java/ru/practicum/ewm/dto/EventShortDto.java
@@ -16,7 +16,7 @@ public class EventShortDto {
 
     private CategoryDto category;
 
-    private int confirmedRequests;
+    private Long confirmedRequests;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
